% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{arcade_revenue}
\alias{arcade_revenue}
\title{'Total revenue generated by arcades' correlates with
 'Computer science doctorates awarded in the US'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.985.
}

