% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{cheese_consumption}
\alias{cheese_consumption}
\title{'Per capita cheese consumption' correlates with
 'Number of people who died by becoming tangled in their bedsheets'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.947.
}

