% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{chicken_consumption}
\alias{chicken_consumption}
\title{'Per capita consumption of chicken' correlates with
 'Total US crude oil imports'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.899.
}

