% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{divorce_rate_maine}
\alias{divorce_rate_maine}
\title{'Divorce rate in Maine' correlates with
 'Per capita consumption of margarine'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.992.
}

