% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{fishing_boat_drownings}
\alias{fishing_boat_drownings}
\title{'People who drowned after falling out of a fishing boat' correlates with
 'Marriage rate in Kentucky'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.952.
}

