% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{japanese_cars_sold}
\alias{japanese_cars_sold}
\title{'Japanese cars sold in the US' correlates with
 'Suicides by crashing of motor vehicle'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.934.
}

