% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{math_doctorates}
\alias{math_doctorates}
\title{'Mathematics doctorates awarded' correlates with
 'Uranium stored at nuclear power plants'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.952.
}

