% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{miss_america_age}
\alias{miss_america_age}
\title{'Age of Miss America' correlates with
 'Murders by steam, hot vapours and hot objects'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.870.
}

