% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{mozzarella_consumption}
\alias{mozzarella_consumption}
\title{'Per capita consumption of mozzarella cheese' correlates with
 'Civil engineering doctorates awarded in the US'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.959.
}

