% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{norway_oil_imports}
\alias{norway_oil_imports}
\title{'US crude oil imports from Norway' correlates with
 'Drivers killed in collision with railway train'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.955.
}

