% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{pool_drownings}
\alias{pool_drownings}
\title{'Number of people who drowned by falling into a pool' correlates with
 'Films Nicolas Cage appeared in'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.666.
}

