% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{space_launches}
\alias{space_launches}
\title{'Worldwide non-commercial space launches' correlates with
 'Sociology doctorates awarded in the US'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.789.
}

