% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{spelling_bee_letters}
\alias{spelling_bee_letters}
\title{'Letters in the winning word of the Scripps National Spelling Bee' correlates with
 'Number of people killed by venomous spiders'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.805.
}

