% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{swimming_pool_drownings}
\alias{swimming_pool_drownings}
\title{'Number of people who drowned while in a swimming-pool' correlates with
 'Power generated by US nuclear plants'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.901.
}

