% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in spuriouscorrelations-package.R
\name{us_spending}
\alias{us_spending}
\title{'US spending on science, space, and technology' correlates with
 'Suicides by hanging, strangulation and suffocation'}
\description{
A dataset to preserve tylervigen.com correlations. The correlation
 is 0.992.
}

